Started adding the 1 and 0s to each Node, will think more about it and look more into huffman to figure out more

+++ CWASM/hoffman_coding/assign_num.h
@@ -4,6 +4,6 @@
 #include "tree.h"
 #include <stddef.h>
 
-int assign_num(Node **nodes, size_t *node_count);
+int assign_num(Node **nodes, int *node_count);
 
 #endif
